fix: investigation data
