+++ services/workflows-service/src/data-analytics/data-analytics.service.ts
@@ -1,26 +1,25 @@
-import { Injectable } from '@nestjs/common';
+import { MERCHANT_REPORT_TYPES_MAP, MerchantReportType } from '@/business-report/constants';
+import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 import { PrismaService } from '@/prisma/prisma.service';
+import { Injectable } from '@nestjs/common';
+import { AlertSeverity, Prisma } from '@prisma/client';
+import { isEmpty } from 'lodash';
+import { AggregateType } from './consts';
 import {
   CheckRiskScoreOptions,
+  DailySingleTransactionAmountType,
   HighTransactionTypePercentage,
   HighVelocityHistoricAverageOptions,
   InlineRule,
   TCustomersTransactionTypeOptions,
   TDormantAccountOptions,
-  TPeerGroupTransactionAverageOptions,
-  TransactionsAgainstDynamicRulesType,
-  TMultipleMerchantsOneCounterparty,
   TExcludedCounterparty,
   TMerchantGroupAverage,
-  DailySingleTransactionAmountType,
+  TMultipleMerchantsOneCounterparty,
+  TPeerGroupTransactionAverageOptions,
+  TransactionsAgainstDynamicRulesType,
 } from './types';
-import { AggregateType, TIME_UNITS } from './consts';
 import { calculateStartDate } from './utils';
-import { Alert, AlertSeverity, PaymentMethod, Prisma, TransactionRecordType } from '@prisma/client';
-import { AppLoggerService } from '@/common/app-logger/app-logger.service';
-import { isEmpty } from 'lodash';
-import { MERCHANT_REPORT_TYPES_MAP, MerchantReportType } from '@/business-report/constants';
-import { SubjectRecord } from '@/alert/types';
 
 const COUNTERPARTY_ORIGINATOR_JOIN_CLAUSE = Prisma.sql`JOIN "Counterparty" AS "cpOriginator" ON "tr"."counterpartyOriginatorId" = "cpOriginator"."id"`;
 const COUNTERPARTY_BENEFICIARY_JOIN_CLAUSE = Prisma.sql`JOIN "Counterparty" AS "cpBeneficiary" ON "tr"."counterpartyBeneficiaryId" = "cpBeneficiary"."id"`;
@@ -302,7 +301,7 @@ export class DataAnalyticsService {
 
     if (amountBetween) {
       conditions.push(
-        Prisma.sql`"transactionAmount" BETWEEN ${amountBetween.min} AND ${amountBetween.max}`,
+        Prisma.sql`"transactionBaseAmount" BETWEEN ${amountBetween.min} AND ${amountBetween.max}`,
       );
     }
 
@@ -389,7 +388,7 @@ export class DataAnalyticsService {
           "${Prisma.raw(subjectColumn)}"
       )
       SELECT
-        "${Prisma.raw(subjectColumn)}" AS "counterpartyId"
+        "${Prisma.raw(subjectColumn)}"
       FROM
         "transactionsData"
       WHERE
@@ -411,10 +410,10 @@ export class DataAnalyticsService {
     const conditions: Prisma.Sql[] = [
       Prisma.sql`"tr"."projectId" = ${projectId}`,
       Prisma.sql`jsonb_exists(config, 'customer_expected_amount') AND ((config ->> 'customer_expected_amount')::numeric * ${factor}) != ${customerExpectedAmount}`,
-      Prisma.sql`"tr"."transactionAmount" > (config ->> 'customer_expected_amount')::numeric`,
+      Prisma.sql`"tr"."transactionBaseAmount" > (config ->> 'customer_expected_amount')::numeric`,
     ];
 
-    const query: Prisma.Sql = Prisma.sql`SELECT "tr"."businessId" , "tr"."transactionAmount" FROM "TransactionRecord" as "tr"
+    const query: Prisma.Sql = Prisma.sql`SELECT "tr"."businessId" , "tr"."transactionBaseAmount" FROM "TransactionRecord" as "tr"
     WHERE ${Prisma.join(conditions, ' AND ')} `;
 
     const results = await this.prisma.$queryRaw(query);
@@ -859,150 +858,5 @@ AND a.activeDaysTransactions > ((a.lastTransactionsCount - a.activeDaysTransacti
     });
 
     return await this.prisma.$queryRaw<T>(query);
-  }
-  private buildTransactionsFiltersByAlert(inlineRule: InlineRule, alert?: Alert) {
-    const whereClause: Prisma.TransactionRecordWhereInput = {};
-
-    const filters: {
-      endDate: Date | undefined;
-      startDate: Date | undefined;
-    } = {
-      endDate: undefined,
-      startDate: undefined,
-    };
-
-    if (alert) {
-      const endDate = alert.updatedAt || alert.createdAt;
-      endDate.setHours(23, 59, 59, 999);
-      filters.endDate = endDate;
-    }
-
-    // @ts-ignore - TODO: Replace logic with proper implementation for each rule
-    // eslint-disable-next-line
-    let { timeAmount, timeUnit } = inlineRule.options;
-
-    if (!timeAmount || !timeUnit) {
-      if (
-        inlineRule.fnName === 'evaluateHighVelocityHistoricAverage' &&
-        inlineRule.options.lastDaysPeriod &&
-        timeUnit
-      ) {
-        timeAmount = inlineRule.options.lastDaysPeriod.timeAmount;
-      } else {
-        return filters;
-      }
-    }
-
-    let startDate = new Date();
-
-    let subtractValue = 0;
-
-    const baseSubstractByMin = timeAmount * 60 * 1000;
-
-    switch (timeUnit) {
-      case TIME_UNITS.minutes:
-        subtractValue = baseSubstractByMin;
-        break;
-      case TIME_UNITS.hours:
-        subtractValue = 60 * baseSubstractByMin;
-        break;
-      case TIME_UNITS.days:
-        subtractValue = 24 * 60 * baseSubstractByMin;
-        break;
-      case TIME_UNITS.months:
-        startDate.setMonth(startDate.getMonth() - timeAmount);
-        break;
-      case TIME_UNITS.years:
-        startDate.setFullYear(startDate.getFullYear() - timeAmount);
-        break;
-    }
-
-    startDate.setHours(0, 0, 0, 0);
-    startDate = new Date(startDate.getTime() - subtractValue);
-
-    if (filters.endDate) {
-      startDate = new Date(Math.min(startDate.getTime(), filters.endDate.getTime()));
-    }
-
-    filters.startDate = startDate;
-
-    if (filters.startDate) {
-      whereClause.transactionDate = {
-        gte: filters.startDate,
-      };
-    }
-
-    if (filters.endDate) {
-      whereClause.transactionDate = {
-        lte: filters.endDate,
-      };
-    }
-
-    return whereClause;
-  }
-
-  getInvestigationFilter(projectId: string, inlineRule: InlineRule, subject: SubjectRecord) {
-    let investigationFilter;
-
-    switch (inlineRule.fnInvestigationName) {
-      case 'investigateTransactionsAgainstDynamicRules':
-        investigationFilter = this[inlineRule.fnInvestigationName]({
-          ...inlineRule.options,
-          projectId,
-        });
-    }
-
-    if (!investigationFilter) {
-      this.logger.error(`No evaluation function found`, {
-        inlineRule,
-      });
-
-      throw new Error(
-        `No evaluation function found for rule name: ${(inlineRule as InlineRule).id}`,
-      );
-    }
-
-    return {
-      ...subject,
-      ...investigationFilter,
-      ...this.buildTransactionsFiltersByAlert(inlineRule),
-      projectId,
-    } satisfies Prisma.TransactionRecordWhereInput;
-  }
-
-  investigateTransactionsAgainstDynamicRules(options: TransactionsAgainstDynamicRulesType) {
-    const {
-      amountBetween,
-      direction,
-      transactionType,
-      paymentMethods = [],
-      excludePaymentMethods = false,
-      projectId,
-    } = options;
-
-    return {
-      projectId,
-      ...(amountBetween
-        ? {
-            transactionAmount: {
-              gte: amountBetween?.min,
-              lte: amountBetween?.max,
-            },
-          }
-        : {}),
-      ...(direction ? { transactionDirection: direction } : {}),
-      ...(transactionType
-        ? {
-            transactionType: {
-              in: transactionType as TransactionRecordType[],
-            },
-          }
-        : {}),
-      paymentMethod: {
-        ...(excludePaymentMethods
-          ? { notIn: paymentMethods as PaymentMethod[] }
-          : { in: paymentMethods as PaymentMethod[] }),
-      },
-    } as const satisfies Prisma.TransactionRecordWhereInput;
   }
 }
